Centralize shared TypeScript types
