+++ src/types/group.ts
@@ -1,9 +1,33 @@
-export type Group = {
+export interface Participant {
   id: string;
   name: string;
-  dayOfWeek: string; // ex: "monday", "tuesday", etc.
-  startTime: string; // ex: "14:00"
-  endTime: string;   // ex: "15:00"
+  avatarUrl?: string;
+  dataAiHint?: string;
+}
+
+export interface GroupResource {
+  id: string;
+  name: string;
+  type: "pdf" | "docx" | "image" | "link" | "other";
+  url?: string;
+  uploadDate: string;
+  description?: string;
+  dataAiHint?: string;
+}
+
+export interface Group {
+  id: string;
+  name: string;
+  psychologist: string;
   psychologistId: string;
+  membersCount: number;
+  schedule: string;
+  dayOfWeek: string;
+  startTime: string;
+  endTime: string;
+  nextSession?: string;
+  description?: string;
+  participants: Participant[];
   meetingAgenda?: string;
-};
+  resources?: GroupResource[];
+}
